gc agent/requirements.txt
streamlit
groq
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
dateparser
pytz
